requirements.txt
appnope==0.1.0
arrow==0.14.6
backcall==0.1.0
chardet==3.0.4
decorator==4.4.0
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
mpu==0.20.0
parso==0.5.1
pdfminer.six==20181108
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycryptodome==3.9.0
Pygments==2.4.2
PyPDF2==1.26.0
python-dateutil==2.8.0
six==1.12.0
sortedcontainers==2.1.0
traitlets==4.3.2
wcwidth==0.1.7
